requirements.txt
# Core dependencies for Trakt2EmbySync
streamlit>=1.22.0
python-dotenv>=0.21.0
requests>=2.28.1
schedule>=1.1.0
watchdog>=2.1.9
